fix(z): Address inconsistent mremap return arguments with arm64 (#309)
This PR adds a specific arm64 implementation for the `mremap` function
which ignores the second return argument from the syscall. **All other
Linux implementations are unchanged.**

This PR reworks [an earlier
PR](https://github.com/dgraph-io/ristretto/pull/281) so that the
functionality on non-arm64 platforms remains the same.

Tested on Linux amd64 and Linux arm64.
